// portal/ui/src/app/interfaces/ICommunication.ts
export interface IDeveloper {
  id: string;
  clients: IDeveloperMapping[];
}

export interface IDeveloperMapping {
  clientId: string;
  organizationId: string;
}

export interface IApi {
  organization: IOrganization;
  id: string;
  name: string;
  description: string;
  createdBy: string;
  image?: string;

  /** @format date-time */
  createdOn: string;

  /** @format int32 */
  numPublished: number;
}

export interface IApiGateway {
  gatewayId: string;
}

export interface IApiPlan {
  planId: string;
  version: string;
  discoverability: Discoverability;
}

export interface IApiVersion {
  /** @format int64 */
  id: number;
  api: IApi;
  status: 'Created' | 'Ready' | 'Published' | 'Retired';
  endpoint: string;
  endpointType: 'rest' | 'soap' | 'ui';
  endpointContentType: 'json' | 'xml';
  endpointProperties: Record<string, string>;
  gateways: IApiGateway[];
  publicAPI: boolean;
  plans: IApiPlan[];
  version: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;

  /** @format date-time */
  publishedOn: string;

  /** @format date-time */
  retiredOn: string;
  definitionType:
    | 'None'
    | 'SwaggerJSON'
    | 'SwaggerYAML'
    | 'WSDL'
    | 'WADL'
    | 'RAML'
    | 'External';
  parsePayload: boolean;
  disableKeysStrip: boolean;
  definitionUrl: string;
  extendedDescription: string;
  publicDiscoverability: Discoverability;
}
export interface IOrganization {
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;
}
// ...
export interface IGateway {
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;
  type: 'REST';
  configuration: string;
}
// ...
export interface IClientSummary {
  organizationId: string;
  organizationName: string;
  id: string;
  name: string;
  description: string;

  /** @format int32 */
  numContracts: number;
}
// ...
export interface IPolicyDefinition {
  id: string;
  policyImpl: string;
  name: string;
  description: string;
  icon: string;
  templates: IPolicyDefinitionTemplate[];

  /** @format int64 */
  pluginId?: number;
  formType: 'Default' | 'JsonSchema';
  form?: string;
  deleted: boolean;
}

export interface IPolicyDefinitionTemplate {
  language: string | null;
  template: string;
}
// ...
export interface IApiSummary {
  organizationId: string;
  organizationName: string;
  id: string;
  name: string;
  description: string;
  image: string;

  /** @format date-time */
  createdOn: string;
}
// ...
export interface IClient {
  organization: IOrganization;
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
}

export interface IClientVersion {
  /** @format int64 */
  id: number;
  client: IClient;
  status:
    | 'Created'
    | 'Ready'
    | 'Registered'
    | 'Retired'
    | 'AwaitingApproval'
    | 'Unregistered';
  version: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;

  /** @format date-time */
  publishedOn: string;

  /** @format date-time */
  retiredOn: string;
  apikey: string;
}
// ...
export interface IPlugin {
  /** @format int64 */
  id: number;
  groupId: string;
  artifactId: string;
  version: string;
  classifier: string;
  type: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  deleted: boolean;
}

export interface INewPlugin {
  groupId: string;
  artifactId: string;
  version: string;
  classifier: string;
  type: string;
  name: string;
  description: string;
  upgrade: boolean;
}

export interface IUpdatePolicyDefinition {
  name: string;
  description: string;
  icon: string;
}

export interface IRole {
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  autoGrant: boolean;
  permissions: (
    | 'orgView'
    | 'orgEdit'
    | 'orgAdmin'
    | 'apiView'
    | 'apiEdit'
    | 'apiAdmin'
    | 'clientView'
    | 'clientEdit'
    | 'clientAdmin'
    | 'planView'
    | 'planEdit'
    | 'planAdmin'
  )[];
}
// ...
export interface IUser {
  username: string;
  fullName: string;
  email: string;

  /** @format date-time */
  joinedOn: string;
  admin: boolean;
}
// ...
export interface IMetaData {
  apimanVersion: string;
  exportedOn: string;
  id: number;
}

export interface IBlob {
  blob: string;
  createdOn: string;
  hash: number;
  id: string;
  mimeType: string;
  modifiedOn: string;
  name: string;
  refCount: number;
}

export interface IApimanData {
  Blobs: IBlob[];
  Developers: IDeveloper[];
  Gateways: IGateway[];
  Metadata: IMetaData;
  Orgs: IOrganization[];
  Plugins: IPlugin[];
  PolicyDefinitions: IPolicyDefinition[];
  Roles: IRole[];
  Users: IUser[];
}

export enum Discoverability {
  PORTAL = 'PORTAL',
  ANONYMOUS = 'ANONYMOUS',
  FULL_PLATFORM_MEMBERS = 'FULL_PLATFORM_MEMBERS',
  ORG_MEMBERS = 'ORG_MEMBERS'
}